Move term table page classes into term_table namespace
These are really representing different kinds of table, not different
kinds of page, so putting them into this namespace makes that clear. The
next thing to change will be to actually make these table level.
